requirements.txt
smolagents[openai]>=1.24.0
pydantic>=2.0
python-dotenv>=1.0
pyyaml
click
sympy
numpy
scipy
networkx
z3-solver
